feat: load files to S3 before each test
