+++ .travis.yml
@@ -13,10 +13,9 @@ install:
   - docker-compose run --rm dev composer install -n
   - docker-compose run --rm dev ./vendor/bin/phpcs --standard=psr2 -n --ignore=vendor --extensions=php .
   - docker-compose run --rm dev ./vendor/bin/phpstan analyse ./src --level=4 -c phpstan.neon
-  - docker-compose run --rm dev php ./tests/loadS3.php
 
 script:
-  - docker-compose run --rm tests
+  - docker-compose run --rm dev ./vendor/bin/phpunit
 
 after_success:
   - docker images
