+++ composer.json
@@ -33,7 +33,8 @@
   },
   "autoload": {
     "psr-4": {
-      "Keboola\\": "src/Keboola/"
+      "Keboola\\": "src/Keboola/",
+      "Keboola\\S3ExtractorTest\\": "tests/Keboola/S3ExtractorTest"
     }
   }
 }
