+++ tests.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
 set -e
 
-php ./tests/loadS3.php
 ./vendor/bin/phpunit
