+++ tests/Keboola/S3ExtractorTest/ApplicationTest.php
@@ -1,34 +1,12 @@
 <?php
+
 namespace Keboola\S3ExtractorTest;
 
 use Keboola\S3Extractor\Application;
-use Keboola\S3Extractor\Exception;
-use Monolog\Handler\NullHandler;
 use Monolog\Handler\TestHandler;
-use PHPUnit\Framework\TestCase;
-use Symfony\Component\Filesystem\Filesystem;
-use Symfony\Component\Process\Process;
 
 class ApplicationTest extends TestCase
 {
-    const AWS_REGION_ENV = 'AWS_REGION';
-    const AWS_S3_BUCKET_ENV = 'AWS_S3_BUCKET';
-    const AWS_S3_ACCESS_KEY_ENV = 'DOWNLOAD_USER_AWS_ACCESS_KEY';
-    const AWS_S3_SECRET_KEY_ENV = 'DOWNLOAD_USER_AWS_SECRET_KEY';
-    protected $path = '/tmp/application';
-
-    public function setUp()
-    {
-        if (!file_exists($this->path)) {
-            mkdir($this->path);
-        }
-    }
-
-    public function tearDown()
-    {
-        passthru('rm -rf ' . $this->path);
-    }
-
     public function testApplication()
     {
         $config = [
