+++ tests/Keboola/S3ExtractorTest/ExceptionsTest.php
@@ -4,29 +4,10 @@ namespace Keboola\S3ExtractorTest;
 
 use Keboola\S3Extractor\Application;
 use Keboola\S3Extractor\Exception;
-use Keboola\S3Extractor\Extractor;
 use Monolog\Handler\TestHandler;
-use PHPUnit\Framework\TestCase;
 
 class ExceptionsTest extends TestCase
 {
-    const AWS_S3_BUCKET_ENV = 'AWS_S3_BUCKET';
-    const AWS_S3_ACCESS_KEY_ENV = 'DOWNLOAD_USER_AWS_ACCESS_KEY';
-    const AWS_S3_SECRET_KEY_ENV = 'DOWNLOAD_USER_AWS_SECRET_KEY';
-    protected $path = '/tmp/errors';
-
-    public function setUp()
-    {
-        if (!file_exists($this->path)) {
-            mkdir($this->path);
-        }
-    }
-
-    public function tearDown()
-    {
-        passthru('rm -rf ' . $this->path);
-    }
-
     public function testInvalidBucket()
     {
         $this->expectException(Exception::class);
