+++ tests/Keboola/S3ExtractorTest/NewFilesOnlyTest.php
@@ -1,36 +1,13 @@
 <?php
+
 namespace Keboola\S3ExtractorTest;
 
 use Keboola\S3Extractor\Extractor;
-use Monolog\Formatter\LineFormatter;
-use Monolog\Handler\StreamHandler;
 use Monolog\Handler\TestHandler;
 use Monolog\Logger;
-use PHPUnit\Framework\TestCase;
 
 class NewFilesOnlyTest extends TestCase
 {
-    const AWS_S3_BUCKET_ENV = 'AWS_S3_BUCKET';
-    const AWS_S3_ACCESS_KEY_ENV = 'DOWNLOAD_USER_AWS_ACCESS_KEY';
-    const AWS_S3_SECRET_KEY_ENV = 'DOWNLOAD_USER_AWS_SECRET_KEY';
-    const UPDATE_AWS_S3_ACCESS_KEY_ENV = 'UPLOAD_USER_AWS_ACCESS_KEY';
-    const UPDATE_AWS_S3_SECRET_KEY_ENV = 'UPLOAD_USER_AWS_SECRET_KEY';
-    const UPDATE_AWS_S3_BUCKET = 'AWS_S3_BUCKET';
-    const UPDATE_AWS_REGION = 'AWS_REGION';
-
-    protected $path;
-
-    public function setUp()
-    {
-        $this->path = '/tmp/aws-s3-extractor/' . uniqid();
-        mkdir($this->path, 0777, true);
-    }
-
-    public function tearDown()
-    {
-        system('rm -rf ' . $this->path);
-    }
-
     public function testSuccessfulDownloadFromRoot()
     {
         $key = "file1.csv";
